Only use .cmt files if they are an interface
Change the behaviour of extract so that it only uses a .cmt file if that
file is an interface file. Otherwise it falls back to the .cmi file.

This meant changing how extractions are represented by codocExtraction,
since we need to keep all possible sources for a file in order to
fallback to the .cmi if the .cmt is not an interface. I also tried to
generally simplify this module as it was very complicated for what it
was doing.

The "non-cmti source" issue has been replaced by a "cmi source" issue,
since using a cmt is correct for files with no .mli, and we no longer
use them for files with an .mli.