do not strip pkg_dir in extract folder
It is not obvious when we really need to strip the `_build/` prefix,
but stripping is definitely made under the assumption that the `pkg_dir`
starts with `_build/`. This assumption might not hold. Issue #94
demonstrates it. To fix #94 it is proposed to apply the stripping only
if `pkg_dir` indeed starts with `_build`.

It is possible, that a better solution would be to strip `_build` if the
pkg_dir starts with `_build` without checking the list of
extractables. Or even don't strip if it is not needed.